[Unity][TVMScript] Fix prim_func lost issue in relax.emit_te (#14189)
Previously R.emit_te was introduced in https://github.com/apache/tvm/pull/14123. The `prim_func`s were not added into the same `ir_module`. The pr is to fix this issue, move some `call_tir` input handling code from `bb.call_te` to utils, then it is able to be leveraged by both `bb.emit_te` and `R.emit_te`.
