+++ python/tvm/relax/block_builder.py
@@ -16,7 +16,6 @@
 # under the License.
 # pylint: disable=no-else-return, invalid-name
 """Developer API of constructing Relax AST."""
-import typing
 
 from typing import Dict, List, Optional, Union, Any, Callable
 from tvm.ir.module import IRModule
@@ -25,18 +24,17 @@ from tvm import relax as rx, tir
 import tvm
 from .expr import (
     Expr,
-    te_tensor,
     Var,
-    ShapeExpr,
     GlobalVar,
     BindingBlock,
     Tuple,
     BaseFunc,
     Binding,
 )
-from .struct_info import PrimStructInfo, ShapeStructInfo, StructInfo, TensorStructInfo
+from .struct_info import StructInfo
 from .op.base import call_tir
 from . import _ffi_api
+from .utils import gen_call_tir_inputs
 
 
 class FunctionScope(object):
@@ -196,107 +194,6 @@ class BlockBuilder(Object):
             if not is_emit_func_output_called:
                 raise RuntimeError("emit_func_output must be called in a relax function.")
 
-    def _convert_te_arg(
-        self, te_args: Any, tir_var_map: Dict[tir.Var, tir.PrimExpr]
-    ) -> typing.Tuple[Any, List[tvm.te.Tensor]]:
-        """Helper function used by `call_te` to convert Relax expressions to TE tensor.
-
-        In the common case, the type of te_args is a Relax expression and is converted
-        into a TE tensor.
-        If te_args is a nested or recursive datatype (i.e list, dict, tvm.ir.Map, tvm.ir.Array),
-        we recursive and convert any value of type Relax expression into a TE tensor.
-        Common values of type int, float, and str are preserved.
-
-        In dynamic shape cases, the passed in arguments may contain TIR variable.
-        For example, the argument can be a Relax Var with TensorStructInfo, which
-        has symbolic shape, or the argument can be a ShapeExpr with symbolic variables.
-        To make the PrimFunc generated by `call_te` has independent variables with
-        the caller Relax function, we will substitute the TIR variables in the input
-        arguments with fresh ones, which is done by maintaining a TIR variable mapping.
-
-        Parameters
-        ----------
-        te_args : Any
-            Argument to convert to TE
-
-        tir_var_map : Dict[tir.Var, tir.PrimExpr]
-            The TIR variable mapping, which maps TIR variables on the Relax function
-            side to the new set of variables used on the PrimFunc side.
-
-        Returns
-        -------
-        ret : (Any, [tvm.te.Tensor])
-            A tuple of the converted te_args, and a list of te tensors for each converted
-            Relax expression
-        """
-        te_args_list = []
-
-        def _copy_undefined_var(expr: tir.PrimExpr):
-            def _visit_expr(e: tir.PrimExpr):
-                if isinstance(e, tir.Var) and e not in tir_var_map:
-                    new_var = tir.Var(e.name, e.dtype)
-                    tir_var_map[e] = new_var
-
-            tir.stmt_functor.post_order_visit(expr, _visit_expr)
-
-        def _convert_te_arg_helper(arg):
-            if isinstance(arg, Expr):  # type: ignore
-                if isinstance(arg.struct_info, TensorStructInfo):
-                    assert isinstance(
-                        arg.struct_info.shape, ShapeExpr
-                    ), "emit_te now only supports Tensor that has ShapeExpr shape"
-                    for shape_value in arg.struct_info.shape.values:
-                        _copy_undefined_var(shape_value)
-
-                    arg = te_tensor(arg, tir_var_map)
-                    te_args_list.append(arg)
-                    return arg
-                elif isinstance(arg.struct_info, ShapeStructInfo):
-                    assert isinstance(
-                        arg, ShapeExpr
-                    ), "For Expr having ShapeStructInfo, emit_te now only supports ShapeExpr"
-                    return [_convert_te_arg_helper(val) for val in arg.values]
-                elif isinstance(arg.struct_info, PrimStructInfo):
-                    return arg.value
-            elif isinstance(arg, (list, tvm.ir.Array)):
-                return [_convert_te_arg_helper(x) for x in arg]
-            elif isinstance(arg, tuple):
-                return tuple([_convert_te_arg_helper(x) for x in arg])
-            elif isinstance(arg, (dict, tvm.ir.Map)):
-                for key in arg:
-                    assert isinstance(
-                        key, str
-                    ), "emit_te only supports dict with string as the key currently"
-                return {k: _convert_te_arg_helper(arg[k]) for k in arg}
-            elif isinstance(arg, tir.PrimExpr):
-                _copy_undefined_var(arg)
-                return tir.stmt_functor.substitute(arg, tir_var_map)
-            elif isinstance(arg, (int, float, str, tvm.ir.Type, tvm.ir.Attrs)) or arg is None:
-                return arg
-            raise TypeError("not supported type in emit_te: {}".format(type(arg)))
-
-        new_arg = _convert_te_arg_helper(te_args)
-        return new_arg, te_args_list
-
-    def _get_unbound_tir_vars(self, args: List[tvm.te.Tensor]) -> List[tvm.tir.Var]:
-        """get unbound TIR vars (i.e TIR vars used in the shape but is not
-        itself a dimension of a shape)"""
-        bound_vars = set()
-        used_vars = set()
-
-        def _populate_used_vars(expr):
-            if isinstance(expr, tvm.tir.Var):
-                used_vars.add(expr)
-
-        for x in args:
-            for s in x.shape:
-                tvm.tir.stmt_functor.post_order_visit(s, _populate_used_vars)
-                if isinstance(s, tir.Var):
-                    bound_vars.add(s)
-
-        diff = used_vars - bound_vars
-        return list(diff)
-
     def function(
         self,
         name: str,
@@ -410,61 +307,13 @@ class BlockBuilder(Object):
             A newly created call node
         """
 
-        primfunc_name_hint = kwargs.pop("primfunc_name_hint", None)
-        tir_var_map: Dict[tir.Var, tir.PrimExpr] = dict()
-        new_args, te_arg_list = self._convert_te_arg(args, tir_var_map)
-        new_kwargs, te_kwarg_list = self._convert_te_arg(kwargs, tir_var_map)
+        primfunc_name = kwargs.pop("primfunc_name_hint", None)
+        tir_func, call_args, output_sinfo, tir_vars = gen_call_tir_inputs(func, *args, **kwargs)
+        if not primfunc_name:
+            primfunc_name = func.__name__
+        gvar = self.add_func(tir_func, primfunc_name)
 
-        te_args = te_arg_list + te_kwarg_list
-
-        te_out = func(*new_args, **new_kwargs)
-        assert isinstance(te_out, tvm.te.tensor.Tensor) or (
-            isinstance(te_out, (tuple, list, tvm.ir.Array))
-            and all(isinstance(t, tvm.te.tensor.Tensor) for t in te_out)
-        ), "only support te.tensor or tuple/list/Array of te.tensor as function output"
-
-        outs = [te_out] if isinstance(te_out, tvm.te.tensor.Tensor) else list(te_out)
-        unbound_tir_vars = self._get_unbound_tir_vars(te_args + outs)
-
-        inputs = [*te_args] + outs
-        tir_func = tvm.te.create_relax_prim_func(inputs, unbound_tir_vars, "int64")
-
-        tir_func = tir_func.without_attr("global_symbol")
-
-        if primfunc_name_hint:
-            gvar = self.add_func(tir_func, primfunc_name_hint)
-        else:
-            gvar = self.add_func(tir_func, func.__name__)
-
-        call_args = [x.op.value for x in te_args]
-
-        def _shape_with_old_tir_var(
-            shape_values: List[tir.PrimExpr], tir_var_inverse_map: Dict[tir.Var, tir.PrimExpr]
-        ):
-            return ShapeExpr(
-                [tir.stmt_functor.substitute(value, tir_var_inverse_map) for value in shape_values]
-            )
-
-        # Invert the TIR variable mapping, to convert the output shape back
-        # with old set of variables.
-        tir_var_inverse_map = {v: k for k, v in tir_var_map.items()}
-
-        output_sinfo = [
-            TensorStructInfo(_shape_with_old_tir_var(out.shape, tir_var_inverse_map), out.dtype)
-            for out in outs
-        ]
-
-        # add arguments for extra parameters from unbound var
-        if len(unbound_tir_vars) > 0:
-            call = call_tir(
-                gvar,
-                call_args,
-                output_sinfo,
-                tir_vars=_shape_with_old_tir_var(unbound_tir_vars, tir_var_inverse_map),
-            )
-        else:
-            call = call_tir(gvar, call_args, output_sinfo)
-        return call
+        return call_tir(gvar, call_args, output_sinfo, tir_vars)
 
     def emit_te(self, func: Callable, *args: Any, **kwargs: Any) -> Var:
         """Emit a call node according to the te function.
